requirements.txt
flask>=2.3
websocket-client>=1.6
